feat: Added Redux utilities to the application
This commit includes a reducer function for the application and wires
up the reducer to the application using Redux and React-Redux. A
container has also been created to bring together Redux and React
in order to manage the state and render the view of the application.

TODO: Implement React-Redux to link up the container with Redux to
inject data into the container for the whole application to start
managing a piece of state using Redux